ゲーム内の数値変換の修正。

+++ DownAction/DownAction/BaseObject.cpp
@@ -12,9 +12,9 @@ bool BaseObject::HitObject(BaseObject* bo1)
 	double a = this->mpPoint->cy + this->mpPoint->gravty + Camera::my;
 	double b = bo1->mpPoint->cy;
 
-	int Yrange = abs(a - b);	// オブジェクト間のY座標総差絶対値
-	int XSrange = (this->mpInfo->sizeX + bo1->mpInfo->sizeX) / 2;		// オブジェクトの横サイズ和値
-	int YSrange = (this->mpInfo->sizeY + bo1->mpInfo->sizeY) / 2;		// オブジェクトの縦サイズ和値
+	int Yrange = (int)abs(a - b);	// オブジェクト間のY座標総差絶対値
+	int XSrange = (int)(this->mpInfo->sizeX + bo1->mpInfo->sizeX) / 2;		// オブジェクトの横サイズ和値
+	int YSrange = (int)(this->mpInfo->sizeY + bo1->mpInfo->sizeY) / 2;		// オブジェクトの縦サイズ和値
 
 	if (Xrange < XSrange && Yrange < YSrange)
 	{
