+++ DownAction/DownAction/ItemChangeScene.h
@@ -35,7 +35,7 @@ private:
 	int nHei;
 	int changeItem;
 	int nowItem[3];
-	const int tp[2][2] = { { 900, 350}, { 900, 750} };// { x, y}
+	const float tp[2][2] = { { 900.0f, 350.0f}, { 900.0f, 750.0f} };// { x, y}
 	bool cFlag;
 	const MRange widR{ 0, 4 };
 	const MRange heiR{ 0, (((int)ItemName::mAll) / widR.max) - 1 };
